improve the performance of the with_scope attributes parsing + use a different name of the simple version of the parser
