+++ lib/locomotive/steam/liquid/tags/path_to.rb
@@ -4,7 +4,7 @@ module Locomotive
       module Tags
         class PathTo < ::Liquid::Tag
 
-          include Concerns::Attributes
+          include Concerns::SimpleAttributesParser
           include Concerns::I18nPage
           include Concerns::Path
